Storybook for player list
