+++ src/reactUi.tsx
@@ -17,7 +17,7 @@ import TitleProvider from './react/TitleProvider'
 import ScoreboardProvider from './react/ScoreboardProvider'
 import SignEditorProvider from './react/SignEditorProvider'
 import IndicatorEffectsProvider from './react/IndicatorEffectsProvider'
-import PlayerListOverlay from './react/PlayerListOverlay'
+import PlayerListOverlayProvider from './react/PlayerListOverlayProvider'
 import SoundMuffler from './react/SoundMuffler'
 import TouchControls from './react/TouchControls'
 import widgets from './react/widgets'
@@ -63,7 +63,7 @@ const InGameUi = () => {
     <RobustPortal to={document.querySelector('#ui-root')}>
       {/* apply scaling */}
       <DeathScreenProvider />
-      <PlayerListOverlay serverIP={'test'} />
+      <PlayerListOverlayProvider serverIP={'test'} />
       <ChatProvider />
       <SoundMuffler />
       <TitleProvider />
